+++ examples/rate-limiting/convex/rateLimiting.ts
@@ -14,6 +14,22 @@ import { v } from "convex/values";
 import { MINUTE, RateLimiter, SECOND } from "@convex-dev/rate-limiter";
 import { DataModel } from "./_generated/dataModel";
 
+const rateLimiter = new RateLimiter(components.rateLimiter, {
+  sendMessage: {
+    kind: "fixed window",
+    period: 5 * SECOND,
+    rate: 1,
+    // Allow accruing usage up to 2 messages to send within 5s (rollover).
+    capacity: 2,
+  },
+  tokenUsage: {
+    kind: "token bucket",
+    period: 1 * MINUTE,
+    rate: 2000,
+    capacity: 10000,
+  },
+});
+
 // Define an agent similarly to the AI SDK
 export const agent = new Agent(components.agent, {
   name: "My First Agent",
@@ -40,17 +56,6 @@ export const agent = new Agent(components.agent, {
       totalTokens: usage.totalTokens,
     });
   },
-});
-
-const rateLimiter = new RateLimiter(components.rateLimiter, {
-  sendMessage: {
-    kind: "fixed window",
-    period: 5 * SECOND,
-    rate: 1,
-    // Allow accruing usage up to 2 messages to send within 5s (rollover).
-    capacity: 2,
-  },
-  tokenUsage: { kind: "token bucket", period: 1 * MINUTE, rate: 1000 },
 });
 
 // This allows us to have a reactive query on the client for when we can send
@@ -81,7 +86,11 @@ export const submitQuestion = mutation({
     await rateLimiter.check(ctx, "tokenUsage", {
       key: userId,
       // TODO: estimate tokens based on all messages, not just the new one.
-      count: await estimateTokens(ctx, args.threadId, args.question),
+      count: Math.max(
+        1,
+        await estimateTokens(ctx, args.threadId, args.question),
+      ),
+      reserve: true,
       throws: true,
     });
 
@@ -103,15 +112,17 @@ export const submitQuestion = mutation({
 // This is a rough estimate of the tokens that will be used.
 // It's not perfect, but it's a good enough estimate for a pre-generation check.
 async function estimateTokens(
-  ctx: MutationCtx,
+  ctx: QueryCtx,
   threadId: string | undefined,
   question: string,
 ) {
-  // Assume roughly 4 characters per token.
+  // Assume roughly 4 characters per token
   const promptTokens = question.length / 4;
+  // Assume 10x as long a reply
+  const estimatedOutputTokens = promptTokens * 3;
   if (!threadId) {
     // This is a new thread, so we guess solely based on the question.
-    return promptTokens;
+    return promptTokens + estimatedOutputTokens;
   }
   const latestMessages = await agent.fetchContextMessages(ctx, {
     threadId,
@@ -130,9 +141,8 @@ async function estimateTokens(
     .reverse()
     .find((message) => message.usage);
   const lastPromptTokens = lastUsageMessage?.usage?.totalTokens ?? 0;
-  return lastPromptTokens + promptTokens;
+  return lastPromptTokens + promptTokens + estimatedOutputTokens;
 }
-
 
 // Step 2: Generate a response to the question. This is a convenience way of
 // making an action that does one generateText LLM call.
@@ -175,5 +185,13 @@ export const recordUsage = internalMutation({
   args: { userId: v.string(), totalTokens: v.number() },
   handler: async (ctx, args) => {
     await ctx.db.insert("usage", args);
+  },
+});
+
+export const getPreviousUsage = query({
+  args: { threadId: v.optional(v.string()) },
+  handler: async (ctx, args) => {
+    // Get usage not accounting for the new question. Do that client-side.
+    return estimateTokens(ctx, args.threadId, "");
   },
 });